✨(site) list the organizations the user can access in the sidebar
Right now, the marsha site is mainly a tool for mass uploading
videos. As such, the sidebar should mainly help the user see at a
glance what organizations they have access to and let them
navigate between them to create playlists and upload videos and
related files.
